refactoring viewmodel actions to dispatcher pattern
