+++ app/src/main/java/com/aleexalvz/cashwise/feature/login/login/LoginUIAction.kt
@@ -1,0 +1,7 @@
+package com.aleexalvz.cashwise.feature.login.login
+
+sealed class LoginUIAction {
+    data class UpdateEmail(val email: String): LoginUIAction()
+    data class UpdatePassword(val password: String): LoginUIAction()
+    data class UpdateRememberMeCheckBox(val rememberMe: Boolean): LoginUIAction()
+}
